+++ web/src/app/chat/documentSidebar/DocumentResults.tsx
@@ -1,6 +1,5 @@
 import { OnyxDocument } from "@/lib/search/interfaces";
 import { ChatDocumentDisplay } from "./ChatDocumentDisplay";
-import { usePopup } from "@/components/admin/connectors/Popup";
 import { removeDuplicateDocs } from "@/lib/documentUtils";
 import { Message } from "../interfaces";
 import {
@@ -11,7 +10,7 @@ import {
   useEffect,
   useState,
 } from "react";
-import { SourcesIcon, XIcon } from "@/components/icons/icons";
+import { XIcon } from "@/components/icons/icons";
 
 interface DocumentResultsProps {
   agenticMessage: boolean;
@@ -53,12 +52,6 @@ export const DocumentResults = forwardRef<HTMLDivElement, DocumentResultsProps>(
     const [delayedSelectedDocumentCount, setDelayedSelectedDocumentCount] =
       useState(0);
 
-    const handleOutsideClick = (event: MouseEvent) => {
-      const sidebar = document.getElementById("onyx-chat-sidebar");
-      if (sidebar && !sidebar.contains(event.target as Node)) {
-        closeSidebar();
-      }
-    };
     useEffect(() => {
       const timer = setTimeout(
         () => {
