+++ web/src/components/chat_search/sources/SourceCard.tsx
@@ -35,7 +35,7 @@ export default function SourceCard({
     <div
       key={doc.document_id}
       onClick={() => openDocument(doc, setPresentingDocument)}
-      className="cursor-pointer text-left overflow-hidden flex flex-col gap-0.5 rounded-lg px-3 py-2 hover:bg-background-dark/80 bg-background-dark/60 w-[200px]"
+      className="cursor-pointer h-[80px] text-left overflow-hidden flex flex-col gap-0.5 rounded-lg px-3 py-2 hover:bg-background-dark/80 bg-background-dark/60 w-[200px]"
     >
       <div className="line-clamp-1 font-semibold text-ellipsis text-text-900  flex h-6 items-center gap-2 text-sm">
         {doc.is_internet || doc.source_type === "web" ? (
@@ -66,31 +66,65 @@ export function SeeMoreBlock({
   uniqueSources,
   toggled,
 }: SeeMoreBlockProps) {
+  // Gather total sources (unique + web).
+  const totalSources = uniqueSources.length + webSourceDomains.length;
+
+  // Filter out "web" from unique sources if we have any webSourceDomains
+  // (preserves the original logic).
   const filteredUniqueSources = uniqueSources.filter(
     (source) => source !== "web" && webSourceDomains.length > 0
   );
-  const numOfWebSourcesToDisplay = 3 - filteredUniqueSources.length;
+
+  // Build a list of up to three icons from the filtered unique sources and web sources.
+  // If we don't reach three icons but have at least one, we'll duplicate the last one.
+  const iconsToRender: Array<{ type: "source" | "web"; data: string }> = [];
+
+  // Push from filtered unique sources (max 3).
+  for (
+    let i = 0;
+    i < filteredUniqueSources.length && iconsToRender.length < 3;
+    i++
+  ) {
+    iconsToRender.push({ type: "source", data: filteredUniqueSources[i] });
+  }
+
+  // Then push from web source domains (until total of 3).
+  for (
+    let i = 0;
+    i < webSourceDomains.length && iconsToRender.length < 3;
+    i++
+  ) {
+    iconsToRender.push({ type: "web", data: webSourceDomains[i] });
+  }
+
+  // If we have fewer than 3 but at least one icon, duplicate the last until we reach 3.
+  while (iconsToRender.length < 3 && iconsToRender.length > 0) {
+    iconsToRender.push(iconsToRender[iconsToRender.length - 1]);
+  }
+
   return (
     <button
       onClick={toggleDocumentSelection}
-      className={`max-w-[260px] min-w-[100px] h-[80px] p-3 bg-[#f1eee8] hover:bg-[#ebe7de] cursor-pointer rounded-lg flex flex-col items-start justify-between transition-opacity duration-300`}
+      className="max-w-[260px] min-w-[150px] h-[80px] p-3 bg-[#f1eee8] hover:bg-[#ebe7de] cursor-pointer rounded-lg flex flex-col items-start justify-between transition-opacity duration-300"
     >
       <div className="flex items-center gap-1">
-        {filteredUniqueSources.slice(0, 3).map((source, index) => (
-          <SourceIcon key={index} sourceType={source} iconSize={14} />
-        ))}
-        {webSourceDomains
-          .slice(0, numOfWebSourcesToDisplay)
-          .map((domain, index) => (
-            <WebResultIcon key={index} url={domain} size={14} />
-          ))}
-        {uniqueSources.length > 3 && (
-          <span className="text-xs text-[#4a4a4a] font-medium ml-1">
-            +{uniqueSources.length - 3}
+        {iconsToRender.map((icon, index) =>
+          icon.type === "source" ? (
+            <SourceIcon
+              key={index}
+              sourceType={icon.data as ValidSources}
+              iconSize={14}
+            />
+          ) : (
+            <WebResultIcon key={index} url={icon.data} size={14} />
+          )
+        )}
+        {totalSources > 3 && (
+          <span className="text-xs text-[#4a4a4a] font-medium">
+            +{totalSources - 3}
           </span>
         )}
       </div>
-
       <div className="text-text-darker text-xs font-semibold">
         {toggled ? "Hide Results" : "Show All"}
       </div>
